refactor: ファイル名の typo hanlder を handler に修正
以下 3 ファイルを git mv でリネームし、参照元の import パスを追随修正した。
シンボル名（エクスポート名等）は変更していない。

- channel_settings_hanlder.ts -> channel_settings_handler.ts
- unique_channel_settings_hanlder.ts -> unique_channel_settings_handler.ts
- unique_role_settings_hanlder.ts -> unique_role_settings_handler.ts
