+++ src/app/handlers/command_handler.ts
@@ -6,13 +6,13 @@ import { exists } from '../common/others';
 import { ErrorTexts } from '../constant/error_texts.js';
 import { handleBan } from '../feat-admin/ban/ban';
 import { channelManagerHandler } from '../feat-admin/channel_manager/channel_manager_handler.js';
-import { channelSettingsHandler } from '../feat-admin/channel_settings/channel_settings_hanlder.js';
+import { channelSettingsHandler } from '../feat-admin/channel_settings/channel_settings_handler.js';
 import { variablesHandler } from '../feat-admin/environment_variables/variables_handler';
 import { festSettingHandler } from '../feat-admin/fest_setting/fest_settings.js';
 import { joinedAtFixer } from '../feat-admin/joined_date_fixer/fix_joined_date.js';
 import { shutdown } from '../feat-admin/shutdown/shutdown_process';
-import { uniqueChannelSettingsHandler } from '../feat-admin/unique_channel_settings/unique_channel_settings_hanlder.js';
-import { uniqueRoleSettingsHandler } from '../feat-admin/unique_role_settings/unique_role_settings_hanlder.js';
+import { uniqueChannelSettingsHandler } from '../feat-admin/unique_channel_settings/unique_channel_settings_handler.js';
+import { uniqueRoleSettingsHandler } from '../feat-admin/unique_role_settings/unique_role_settings_handler.js';
 import { closeCommand } from '../feat-recruit/close_recruit/close_command.js';
 import { anarchyRecruit } from '../feat-recruit/create_recruit/anarchy_recruit.js';
 import { buttonRecruit } from '../feat-recruit/create_recruit/button_recruit.js';
